# Optimize by x50-x80 performance of sortByMarketcap