… src/currencies/__snapshots__/sortByMarketcap.test.js.snap