```
4949 full total currencies & tokens
1831 listed total currencies & tokens
2447 tickers
sortByMarketcapV1 full x 4.93 ops/sec ±32.09% (19 runs sampled)
sortByMarketcapV2 full x 512 ops/sec ±13.50% (58 runs sampled)
sortByMarketcapV1 listed x 23.18 ops/sec ±13.37% (45 runs sampled)
sortByMarketcapV2 listed x 1,565 ops/sec ±2.88% (83 runs sampled)
```